Completed assignment from yesterday divided classes into new files
