+++ author.rb
@@ -1,0 +1,11 @@
+class Author < Person
+
+  attr_reader :name, :gender, :age
+
+  def initialize(args)
+    super
+    @name   = "Bennett"
+    @gender = "male"
+    @age    = 34
+  end
+end
