+++ groceries.txt
@@ -1,0 +1,1 @@
+beer, bologna, batter, butter, beets
